修改ycm_core.so自动化与PluginInstall从命令行执行

+++ init.sh
@@ -26,10 +26,15 @@ if  [ ! -d ~/.vim/bundle/Vundle.vim ] ; then
     git clone https://github.com/VundleVim/Vundle.vim.git ~/.vim/bundle/Vundle.vim
 fi
 
+
+
 #-------------------------------------------------------------------------------
-# copy vimrc to home dir
+# copy vimrc to home dir, and install from command line
 #-------------------------------------------------------------------------------
 cp ./vimrc ~/.vimrc
+set shell=/bin/bash
+vim   +PluginInstall +qall
+
 
 if [ ! -d ~/.vim/bundle/YouCompleteMe ] ; then
     echo "You should do :PluginInstall first"
@@ -41,16 +46,17 @@ fi
 # compile ycm_core
 #-------------------------------------------------------------------------------
 
-cd ~
-mkdir ycm_build
-cd ycm_build
-wget http://releases.llvm.org/5.0.0/clang+llvm-5.0.0-x86_64-linux-gnu-debian8.tar.xz
-tar xJvf clang+llvm-5.0.0-x86_64-linux-gnu-debian8.tar.xz
-cmake -G "Unix Makefiles"  -DPATH_TO_LLVM_ROOT=./clang+llvm-5.0.0-x86_64-linux-gnu-debian8    . ~/.vim/bundle/YouCompleteMe/third_party/ycmd/cpp
-cmake --build . --target ycm_core --config Release
-cd ..
-rm -rf ycm_build
-
+if [ ! -f ~/.vim/bundle/YouCompleteMe/third_party/ycmd/ycm_core.so ] ; then
+    cd ~
+    mkdir ycm_build
+    cd ycm_build
+    wget http://releases.llvm.org/5.0.0/clang+llvm-5.0.0-x86_64-linux-gnu-debian8.tar.xz
+    tar xJvf clang+llvm-5.0.0-x86_64-linux-gnu-debian8.tar.xz
+    cmake -G "Unix Makefiles"  -DPATH_TO_LLVM_ROOT=./clang+llvm-5.0.0-x86_64-linux-gnu-debian8    . ~/.vim/bundle/YouCompleteMe/third_party/ycmd/cpp
+    cmake --build . --target ycm_core --config Release
+    cd ..
+    rm -rf ycm_build
+fi
 
 #-------------------------------------------------------------------------------
 # get the ycm conf
@@ -58,13 +64,3 @@ rm -rf ycm_build
 wget https://raw.githubusercontent.com/zhouzheng12/newycm_extra_conf.py/master/ycm.c.py 
 mkdir -p ~/.vim/ycmconf
 cp ycm.c.py ~/.vim/ycmconf/ycm.c.py
-
-
-
-
-
-
-
-
-
-
